Initial version of gitgroups
Groups are stored in text files in a Git repository. Each text
file lists members' Gerrit username or verified email address:

  $ mkdir Groups
  $ cd Groups
  $ git init
  $ cat >Administrators
  # People who manage the site
  Shawn Pearce <sop@google.com>
  bob@example.com
  that-guy
  ^D
  $ git add Administrators
  $ git commit -m 'Add initial admins'

UUIDs to reference these groups use the git: UUID prefix and
have a simple syntax:

  git:PROJECT:BRANCH:FILE

In the above example we could reference the group as:

  git:Groups:master:Administrators
  git:Groups:Administrators

The HEAD reference is assumed if the BRANCH part is not given,
allowing for shorter UUIDs. The UUID format implies ":" may not
appear in the Git project name or the file name for a group, even
though it is otherwise a legal character in those values.

Change-Id: I328e29481bfd64f8debddee5d352279876bf0d9e